feat(cyclesledger): Generate Rust types (#4023)
# Motivation
We would like to auto-generate the bindings used when calling other
canisters. We are doing this a step at a time, starting with the cycles
ledger, which is an easy choice as it doesn't change much.

# Changes
* Generate types for the cycles ledger client.
* The types have been removed from the manually generated and modified
client.

# Tests
The binding command is run as part of CI, so we will see if the bindings
are out of date.
