+++ scripts/bind/rust/types.hbs
@@ -1,0 +1,15 @@
+//! Bindings to the `{{ canister_name }}` canister, generated by `./scripts/bind/rust.sh`
+//!
+//! Binding configuration: `./scripts/bind/rust/{{ canister_name }}.types.toml`
+//! 
+//! Adapted from: <https://github.com/dfinity/candid/blob/master/rust/candid_parser/src/bindings/rust_call.hbs>
+#![allow(dead_code, unused_imports, clippy::all, clippy::missing_errors_doc)]
+use {{candid_crate}}::{self, CandidType, Deserialize, Principal};
+use ic_cdk::api::call::CallResult;
+
+{{#if submodule}}
+/// Manually maintained methods
+pub mod {{submodule}};
+{{/if}}
+
+{{type_defs}}
