+++ src/cycles_ledger/client/Cargo.toml
@@ -6,5 +6,6 @@ edition = "2021"
 [dependencies]
 candid = { workspace = true }
 ic-cdk = { workspace = true }
+ic-cycles-ledger-types = { workspace = true }
 serde = { workspace = true }
 serde_bytes = { workspace = true }
